Move function.
