--- packages/story-editor/src/utils/checkResourceAccess.js
@@ -1,43 +1,0 @@
-/*
- * Copyright 2021 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Check if the resource can be accessed directly.
- *
- * Makes a HEAD request, which in turn triggers a CORS preflight request
- * in the browser.
- *
- * If the request passes, we don't need to do anything.
- * If it doesn't, it means we need to run the resource through our CORS proxy at all times.
- *
- * @type {function(): boolean}
- */
-async function checkResourceAccess(link) {
-  let shouldProxy = false;
-  if (!link) {
-    return shouldProxy;
-  }
-  try {
-    await fetch(link, {
-      method: 'HEAD',
-    });
-  } catch (err) {
-    shouldProxy = true;
-  }
-
-  return shouldProxy;
-}
-export default checkResourceAccess;
